[README] fix broken links to documentation
There are some broken links to documentation in the top-level README file.
Three of them can be directly replaced with updated URLs, but I don't see
a direct equivalent of "Diem Protocol: Key Concepts". This replaces that with
a "Basic Concepts" link to the start of the "Basics" section in the developer
guides.

Fixes: #9803
Closes: #9838