[consensus] add ability to cancel the timeout task if moving to next round
This commit implements the funtionality of cancelling the timeout task to guarantee there's only one timeout task at the same time.

Previously we fire and forget the timeout event and ignore the stale timeout, but it's error prone.

In one of the recent change (https://github.com/diem/diem/pull/9732), it accidentally switch the channel to diem channel with size 1 so that it's possible
for a stale timeout to override new timeout and stuck the node from making progress.

Closes: #9866