[operational tooling] ported to use rest
json rpc is deprecated and on the way out

this was a rather unfortunate refactor due to use of rest which is only
built on tokio and the operational tool was not async. Therefore it
required porting a lot of the operational tool to async so that it could
then be used in async tests. Otherwise we get the dreaded tokio within
tokio error.