[docker] tweak validator-testnet startup script
With the new faucet, the faucet is expected to immediately perform an
operation on the testnet. It might not be ready, so we spin a few times
to see if the validator is ready before giving up. Might be worth it to
see if it is possible to make one docker container fail if another does,
and just make the faucet run in an infinite loop.

Closes: #385