[storage] VerifiedStateView  gets persisted checkpoint from reader
Instead of requiring it as a parameter to the constructor.

So that it's easier to make persisting state checkpoints asynchronous --
the execution and comission threads don't need to track explicitly the
latest persisted state checkpoint.

We do still need to make sure all speculative SMTs haven't yet been
dropped after the latest persisted checkpoint -- this unfortunately
becomes implicit after this change.

Closes: #1180