[python-sdk] demonstrate bcs multiagent
MultiAgent is particularly painful because it introduces a new signing
object in an attempt to stay compatible with existing transaction
formats. Perhaps it would have been more sensisble for the transaction
to actually take the account addresses of the secondary senders rather
than that be in the authenticator, but the time to fix this has long
since passed.

Data was taken from the following:
https://gist.github.com/davidiw/f3804855caf39081780952838f9e04ee