[framework] add a transfer and optionally create account function
Other blockchains typically allow folks to create accounts and transfer
funds at the same time, because the idea of an account is just ownership
of a private key. Aptos uses a different model, which means that
accounts must be created first.

We could generalize this to any coin or other resource types, but I'm
cautious to support this more generally as each interaction likely has
certain specializers that expect the signer to explicitly opt-in --
except for this one.