[cli][consensus] Add analyze-validator-performance to node cli (#2364)
Adding utilities to analyze validator performance, based on accessing live on-chain data

Fastest to run a full node locally, and connecting cli to it.

Currently fetching from all transactions. We can fetch create API for just metadata transactions and/or fetch from BigQuery, if this is too slow for people. When we run loadtest, and BlockMetadata transactions are not the majority, it does slow down.

Test Plan:
cargo run -p aptos -- node analyze-validator-performance --analyze-mode=all