[consensus] propagate reconfig timestamp to all reconfig suffix blocks
We fixed a race condition in https://github.com/aptos-labs/aptos-core/pull/2082/commits/b3083b202fcdbf0133cc4a1116bfe429a78bf451
which didn't take into consideration of the timestamps for suffix batches.

Previously the timestamp is only fixed for the blocks in the same batch as reconfig block because we reject any suffix batches.
This commit adds an epoch end timestamp and ensures it's applied for all suffix batches.