[nextest] Add retries and don fail fast
Its an unfortunate reality of life that unittests can be flaky, especially when
they're creating databases or interacting with files or both!

To deal with this reality lets not fail PR / main when we have a flaky test,
however in tandem we need to monitor the flaky test in main and work to remove
the flakiness.

Test Plan: there is not a great way to test workflow changes at the moment