[aptos-rosetta] Allow for synthetic blocks
These synthetic blocks allow for a fixed size block, which can flatten
out the amount of transactions in a block request.  However, it has
the caveat that block sizes can be high or low, and may not be as
large as the actual blocks being committed.  This could lead to
different behavior based on the current performance of the network.

Additionally, if you were indexing on this method, it will not be
able to use a different size, without starting over from scratch.
If no block size is given, then it will use the real blocks.