[mempool] remove sequence_number_cache and instead store directly in transaction_store (#4053)
### Description

There were two issues previously:
1. An account's sequence_number_cache could get GC'd before its transactions in mempool were sent to consensus. In this case because a cached sequence number does not exist, the transactions are never sent to consensus (get_batch checks the sequence number).
2. The sequence_number_cache capacity is the same as the mempool, but it could have grown beyond this because sequence_number_cache entries are retained for accounts after transaction removal (until they expire). In this case old sequence_number_cache entries are removed and those transactions may never be sent to consensus (same reason as above).

There doesn't seem to be a need for a separate sequence_number_cache at all. When transactions are submitted to mempool, the latest account sequence number from state db is passed in. The argument is that this sequence number should never be stale compared to notifications coming in to  mempool, as state db will be updated before notifications.

Thus we remove the problematic cache and simply store in the transaction_store. If there are no more transactions, the value is removed (same as the main transaction hashmap after #4041

### Test Plan

Add test for case 2 that would fail in main. Otherwise rely on existing tests.