[forge] Implement ability to run multiple tests again
Summary:

Implement the ability to have multiple tests run in parallel balanced across clusters.

This also introduces a lot of config management for test runs and clusters so
we can spin up / spin down cluster and enable and even run disabled tests but
not care about their outcomes.

This change was landed befoer but then reverted.

Last time there was a change in the run-forge workflow itself which removed the
default suite, this caused the pre forge to fail with "no tests to run" sigh...
This is kind of a result of two things

1) Pre forge step which just leaves a comment is this weird workflow which we
only care about producing enough info to log a comment. This should just be
pushed down into forge.py itself which is more than capable of commenting on a
PR before running

2) run-forge.yaml is not really testable on a PR so a PR which completely
breaks this can actually land. I think I have a solution to this using a pull
request / pull request target depending on if you modify the workflow or not

For context pull request target is used so that we can run build / test on
community PR from forked repos once we've approved a PR. However I think there
is a compromise here. Community members really should not be making changes to
the workflows file themselves so we can use pull request on that and pull
request target on everything else. In the strange case community make a PR for
workflow we can manually test that, but frankly we should have to manually
review that anyway since its a sensitive area.

This reverts commit 5a34576e4f6b9326ad08bc04717015dcb03db542.

Test Plan:

Running locally, running on PR ensure the test suite functinoality works and pr
and continuous work