[gha] Conditionally run pull request and pull request target based on files changed
As a follow up to last week breakage, this attempts to run pull request and
pull request target to make sure that changes to a workflow using pull request
target are both forwards and backwards compatible.

For context, right now if you modify a workflow with pull reqeust target it
will not actually test the workflow on PR so you can break it completely and
the PR would just happily land immediately breaking all workflows!!!!

The longer term that would be ideal is getting rid of pull reqeust target
because currently its impossible to version.

Test Plan: ensure pull request running and passes on PR