[backup] fix flaky test utils::stream::futures_unordered_x::tests
In one of the previous tries I intened to make sure every task yields
but unfortunately I changed the wrong number (actually made sure there's more than 0 tasks, but they can sleep for 0ms).

This time around I'll just wait for the maximum concurrency to be
reached in the tasks.