[Quorum Store] Allow batches even if a validator restarts with a fresh DB (#6364)
# Description

Introduce a nonce into BatchId. If a validator restarts with a fresh DB, the change in nonce indicates that the id will start back at 0, and other validators will accept the new batches.

Without this, a validator with a wiped DB would timeout all batches until it reached the previous batch ID.

# Test Plan

Add smoke test for both when DB exists on restart and when it is wiped on restart. The latter failed before this change.