[abigen] don't panic if you can't handle runtime generics
The current abigen depends on TypeTags to represent abis. Unfortunately
this excludes it from representing any generic entry function. That
should NOT prevent us from writing these entry functions in Move or in
the framework. The current code creates that unnecessary restriction, so
instead of panicing, just return none as if it cannot understand it.