[model-compiler] Use patterns everywhere in the AST; remove interpreter and simplifier
This replaces the use of `LocalVarDecl` accross the model with the new `Pattern` type. This applies to `Block`, `Quant`, and 'Lambda` expression variants.

It also removes some code which is currently not used in applications, namely the stackless bytecode interpreter, the spec simplifier and the flattener. While we are refactoring, we have better agility without this code. We can bring it back from git history later if needed.