[move-examples] token collection offers
Provides the ability to make collection offers to both Tokenv1 and Tokenv2 collections.
A collection offer allows an entity to buy up to N assets within a collection at their
specified amount. The amount offered is extracted from their account and stored at an
escrow. A seller can then exchange the token for the escrowed payment. If it is a
a tokenv2 or the recipient has enabled direct deposit, the token is immediately
transferred. If it is tokenv1 without direct deposit, it is stored in a container
until the recipient extracts it.